Moving SOLID_IDENTITY_PROVIDER and oidcIssuer to contants and updated tests
