+++ src/constants/index.js
@@ -1,4 +1,5 @@
 import RDF_PREDICATES from './rdf_predicates';
 import INTERACTION_TYPES from './interaction_types';
+import { SOLID_IDENTITY_PROVIDER, currentOidcIssuer } from './solid_provider';
 
-export { RDF_PREDICATES, INTERACTION_TYPES };
+export { RDF_PREDICATES, INTERACTION_TYPES, SOLID_IDENTITY_PROVIDER, currentOidcIssuer };
